refactor bulk password scripts to be normalized, use shared lib
